+++ client/lib/views/ViewDeactivate.vue
@@ -88,7 +88,6 @@ async function executeAction({ target, pks }) {
 
 const instanceList = useList({
     props: instanceListProps,
-    paged: false,
     handlers: {
         executeAction,
     },